pyproject.toml
[project]
name = "common-precommits"
version = "0.1.0"
description = "base precommit and associated files for projects"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "git-python>=1.0.3",
    "pre-commit>=4.3.0",
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
]
